requirements.txt
sqlalchemy
psycopg2-binary
pydantic
pydantic[dotenv]
